Add files via upload
The Smart Energy Managment System "SEMS" created in MatLab AppDesigner is the basis for the Nanogrid model created in MatLab Simulink. A smart algorithm applied to the system consisting of renewables, traditional energy sources, a battery (power cloud) and loads allow an efficient management of ressources. The simulation results show a reduced CO2 emittance while energy costs can be saved.